Terry

+++ Projectiles/MutantBoss/MutantBoss.cs
@@ -22,7 +22,7 @@ namespace FargowiltasSouls.Projectiles.MutantBoss
         {
             DisplayName.SetDefault("Mutant");
             Main.projFrames[projectile.type] = 4;
-            ProjectileID.Sets.TrailCacheLength[projectile.type] = 6;
+            ProjectileID.Sets.TrailCacheLength[projectile.type] = 8;
             ProjectileID.Sets.TrailingMode[projectile.type] = 2;
         }
 
@@ -50,10 +50,10 @@ namespace FargowiltasSouls.Projectiles.MutantBoss
                 projectile.Center = Main.npc[ai1].Center;
                 projectile.alpha = Main.npc[ai1].alpha;
                 projectile.direction = projectile.spriteDirection = Main.npc[ai1].direction;
-                projectile.timeLeft = 2;
+                projectile.timeLeft = 30;
                 auraTrail = DisplayAura(Main.npc[ai1]);
 
-                switch((int)Main.npc[ai1].ai[0]) //draw behind whenever holding a weapon
+                /*switch((int)Main.npc[ai1].ai[0]) //draw behind whenever holding a weapon
                 {
                     case 0:
                     case 4:
@@ -74,9 +74,16 @@ namespace FargowiltasSouls.Projectiles.MutantBoss
                     default:
                         projectile.hide = false;
                         break;
-                }
+                }*/
 
-                projectile.frame = (int)(Main.npc[ai1].frame.Y / (float)(Main.projectileTexture[projectile.type].Height / Main.projFrames[projectile.type]));
+                projectile.hide =
+                    Main.player[projectile.owner].ownedProjectileCounts[ModContent.ProjectileType<MutantSpearAim>()] > 0
+                    || Main.player[projectile.owner].ownedProjectileCounts[ModContent.ProjectileType<MutantSpearDash>()] > 0
+                    || Main.player[projectile.owner].ownedProjectileCounts[ModContent.ProjectileType<MutantSpearSpin>()] > 0
+                    || Main.player[projectile.owner].ownedProjectileCounts[ModContent.ProjectileType<MutantSlimeRain>()] > 0;
+
+                if (!Main.dedServ)
+                    projectile.frame = (int)(Main.npc[ai1].frame.Y / (float)(Main.projectileTexture[projectile.type].Height / Main.projFrames[projectile.type]));
 
                 if (Main.npc[ai1].frameCounter == 0)
                 {
@@ -87,9 +94,17 @@ namespace FargowiltasSouls.Projectiles.MutantBoss
             }
             else
             {
-                projectile.Kill();
+                if (Main.netMode != NetmodeID.MultiplayerClient)
+                    projectile.Kill();
                 return;
             }
+        }
+
+        public override void Kill(int timeLeft)
+        {
+            /*Main.NewText("i die now");
+            if (Main.netMode == NetmodeID.Server)
+                NetMessage.BroadcastChatMessage(Terraria.Localization.NetworkText.FromLiteral("i die now aaaaaa"), Color.LimeGreen);*/
         }
 
         public bool DisplayAura(NPC npc)
@@ -121,28 +136,32 @@ namespace FargowiltasSouls.Projectiles.MutantBoss
             int lightningY = lightningFrameHeight * (int)projectile.localAI[0];
             Rectangle lightningRectangle = new Rectangle(0, lightningY, lightning.Width, lightningFrameHeight);*/
 
-            Color color26 = lightColor;
-            color26 = projectile.GetAlpha(color26);
+            Color color26 = projectile.GetAlpha(projectile.hide && Main.netMode == NetmodeID.MultiplayerClient ? Lighting.GetColor((int)projectile.Center.X / 16, (int)projectile.Center.Y / 16) : lightColor);
 
             SpriteEffects effects = projectile.spriteDirection < 0 ? SpriteEffects.None : SpriteEffects.FlipHorizontally;
 
             if (auraTrail)
             {
-                float scale = (Main.mouseTextColor / 200f - 0.35f) * 0.4f + 0.8f;
+                float scale = (Main.mouseTextColor / 200f - 0.35f) * 0.4f + 0.9f;
                 scale *= projectile.scale;
 
                 Main.spriteBatch.Draw(texture2D14, projectile.Center - Main.screenPosition + new Vector2(0f, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(rectangle), Color.White * projectile.Opacity, projectile.rotation, origin2, scale, effects, 0f);
 
-                for (int i = 1; i < ProjectileID.Sets.TrailCacheLength[projectile.type]; i++)
+                for (float i = 1; i < ProjectileID.Sets.TrailCacheLength[projectile.type]; i += 0.5f)
                 {
-                    Color color27 = Color.White * projectile.Opacity * 0.6f;
+                    Color color27 = Color.White * projectile.Opacity * 0.75f;
                     color27 *= (float)(ProjectileID.Sets.TrailCacheLength[projectile.type] - i) / ProjectileID.Sets.TrailCacheLength[projectile.type];
-                    Vector2 value4 = projectile.oldPos[i];
-                    float num165 = projectile.oldRot[i];
-                    Main.spriteBatch.Draw(texture2D14, value4 + projectile.Size / 2f - Main.screenPosition + new Vector2(0, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(rectangle), color27, num165, origin2, projectile.scale, effects, 0f);
+                    int max0 = (int)i - 1;//Math.Max((int)i - 1, 0);
+                    if (max0 < 0)
+                        continue;
+                    Vector2 value4 = projectile.oldPos[max0];
+                    float num165 = projectile.oldRot[max0];
+                    Vector2 center = Vector2.Lerp(projectile.oldPos[(int)i], projectile.oldPos[max0], 1 - i % 1);
+                    center += projectile.Size / 2;
+                    Main.spriteBatch.Draw(texture2D14, center - Main.screenPosition + new Vector2(0, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(rectangle), color27, num165, origin2, projectile.scale, effects, 0f);
                 }
 
-                Main.spriteBatch.Draw(aura, -16 * Vector2.UnitY + projectile.Center - Main.screenPosition + new Vector2(0f, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(auraRectangle), Color.White * projectile.Opacity, projectile.rotation, auraRectangle.Size() / 2f, projectile.scale, effects, 0f);
+                Main.spriteBatch.Draw(aura, -16 * Vector2.UnitY + projectile.Center - Main.screenPosition + new Vector2(0f, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(auraRectangle), Color.White * projectile.Opacity, projectile.rotation, auraRectangle.Size() / 2f, scale, effects, 0f);
             }
             else
             {
@@ -156,7 +175,7 @@ namespace FargowiltasSouls.Projectiles.MutantBoss
                 }
             }
 
-            Main.spriteBatch.Draw(texture2D13, projectile.Center - Main.screenPosition + new Vector2(0f, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(rectangle), projectile.GetAlpha(lightColor), projectile.rotation, origin2, projectile.scale, effects, 0f);
+            Main.spriteBatch.Draw(texture2D13, projectile.Center - Main.screenPosition + new Vector2(0f, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(rectangle), color26, projectile.rotation, origin2, projectile.scale, effects, 0f);
 
             //if (auraTrail) Main.spriteBatch.Draw(lightning, projectile.Center - Main.screenPosition + new Vector2(0f, projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(lightningRectangle), Color.White * projectile.Opacity, projectile.rotation, lightningRectangle.Size() / 2f, projectile.scale, SpriteEffects.None, 0f);
             return false;
